requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
opencv-python
pydantic
reportlab
google-generativeai
python-multipart
python-dotenv